Correcciones en crear_tabla_prenda y crear_tabla_prenda_pal_clave
Se cambió la columna pre_precio_sugerido de tinyInteger a unsignedInteger en la mig crear_tabla_prenda.
Se agregó el comportamiento de onDelete() a las foreign key de la mig crear_tabla_prenda_pal_clave.

+++ database/migrations/2019_01_12_212652_crear_tabla_prenda.php
@@ -23,7 +23,7 @@ class CrearTablaPrenda extends Migration
           $table->string('pre_descripcion', 255)->nullable();
           $table->char('pre_talla', 1)->default('M');
           $table->tinyInteger('pre_cantidad');
-          $table->tinyInteger('pre_precio_sugerido')->nullable();
+          $table->unsignedInteger('pre_precio_sugerido')->nullable();
           $table->date('pre_fecha_compra');
 
           $table->foreign('pre_fk_categoria')
